+++ regression/reference/cxxlibrary/wrapcxxlibrary.h
@@ -32,14 +32,25 @@
 extern "C" {
 #endif
 
+#ifdef __cplusplus
+}
+#endif
+#ifdef __cplusplus
+using CXX_nested = nested;
+#else  // __cplusplus
 
 typedef struct s_CXX_nested CXX_nested;
 struct s_CXX_nested {
+    int index;
     int sublevels;
     CXX_nested * parent;
     CXX_nested * * child;
-    char * name;
 };
+#endif  // __cplusplus
+
+#ifdef __cplusplus
+extern "C" {
+#endif
 
 // splicer begin C_declarations
 // splicer end C_declarations
@@ -59,6 +70,16 @@ void CXX_getGroupName_int32_t_bufferify(int32_t idx, char *SHC_rv,
 
 void CXX_getGroupName_int64_t_bufferify(int64_t idx, char *SHC_rv,
     int SHT_rv_len);
+
+void CXX_nested_get_parent(CXX_nested * SH_this,
+    CXX_SHROUD_array *SHT_rv_cdesc);
+
+void CXX_nested_set_parent(CXX_nested * SH_this, CXX_nested * val);
+
+void CXX_nested_get_child_bufferify(CXX_nested * SH_this,
+    CXX_SHROUD_array *SHT_rv_cdesc);
+
+void CXX_nested_set_child(CXX_nested * SH_this, CXX_nested * * val);
 
 #ifdef __cplusplus
 }