+++ regression/reference/typedefs-cxx/wraptypedefs.cpp
@@ -50,9 +50,7 @@ TYP_TypeID TYP_typefunc(TYP_TypeID arg)
 void TYP_typestruct(TYP_Struct1Rename * arg1)
 {
     // splicer begin function.typestruct
-    Struct1Rename * SHCXX_arg1 = static_cast<Struct1Rename *>
-        (static_cast<void *>(arg1));
-    typestruct(SHCXX_arg1);
+    typestruct(arg1);
     // splicer end function.typestruct
 }
 // end TYP_typestruct